// packages/youtube-mcp/src/api/types.ts
export type YouTubeApiChannelItem = {
  kind?: string;
  etag?: string;
  id: string;
  snippet?: {
    title?: string;
    description?: string;
    customUrl?: string;
    publishedAt?: string;
    thumbnails?: Record<
      string,
      { url: string; width?: number; height?: number }
    >;
    country?: string;
  };
  statistics?: {
    viewCount?: string;
    subscriberCount?: string;
    hiddenSubscriberCount?: boolean;
    videoCount?: string;
  };
  contentDetails?: {
    relatedPlaylists?: {
      likes?: string;
      uploads?: string;
    };
  };
};
// ...
export type YouTubeApiSearchItem = {
  kind?: string;
  etag?: string;
  id:
    | string
    | {
        kind?: string;
        videoId?: string;
        channelId?: string;
        playlistId?: string;
      };
  snippet?: {
    title?: string;
    description?: string;
    channelId?: string;
    channelTitle?: string;
    publishedAt?: string;
    thumbnails?: Record<
      string,
      { url: string; width?: number; height?: number }
    >;
    channelHandle?: string;
    liveBroadcastContent?: string;
  };
};
// ...
export type YouTubeApiResponse<T> = {
  kind?: string;
  etag?: string;
  nextPageToken?: string;
  prevPageToken?: string;
  pageInfo?: {
    totalResults?: number;
    resultsPerPage?: number;
  };
  items?: T[];
};
// ...
export type ChannelDetails = {
  id: string;
  title: string;
  description: string;
  customUrl?: string;
  publishedAt: string;
  thumbnails: Record<string, { url: string; width?: number; height?: number }>;
  viewCount: string;
  subscriberCount: string;
  videoCount: string;
  uploads?: string;
};
// ...
export type SearchResult = {
  id: string;
  title: string;
  description: string;
  channelId: string;
  channelTitle: string;
  publishedAt: string;
  thumbnails: Record<string, { url: string; width?: number; height?: number }>;
  type: "video" | "channel" | "playlist";
};